+++ lib_ksgPlayer/src/main/java/com/ksg/ksgplayer/player/KsgPlayerProxy.java
@@ -361,7 +361,7 @@ public final class KsgPlayerProxy implements IKsgPlayer {
      * @param msc 在指定的位置播放
      */
     @Override
-    public void seekTo(int msc) {
+    public void seekTo(long msc) {
         if (isPlayerAvailable()) {
             this.mInternalPlayer.seekTo(msc);
         }
